requirements.txt
# fastapi==0.111.0
fastapi==0.115.6
alembic==1.13.2
aiohttp==3.10.9
aiosqlite==0.20.0
# sqlmodel==0.0.21
sqlmodel==0.0.22
aiofiles==24.1.0
uvicorn==0.30.1
pydantic==2.8.0
pydantic-settings==2.0.0
pydantic_core==2.20.0
python-multipart==0.0.9

numpy==1.26.3
pandas==2.2.0
tqdm==4.66.1
openpyxl==3.1.2

ruff==0.5.5
pyright==1.1.373
pytest==8.3.2
pytest-asyncio==0.24.0

boto3==1.35.36
botocore==1.35.36
wrapt==1.16.0
aioboto3==13.2.0
aiobotocore==2.15.2
aioitertools==0.12.0

python-ulid==3.0.0
python-jose==3.3.0
asyncpg==0.30.0
passlib[bcrypt]==1.7.4